Update struct.h
integer for room rate

+++ struct.h
@@ -1,7 +1,7 @@
 struct Tenant {
     char username[50];
     char roomNumber[7];
-    float roomRate;
+    int roomRate;
     float waterBill;
     float electricityBill;
     int daysToPay;
